pyproject.toml
[project]
name = "edgedetectionfilter"
version = "0.1.0"
description = ""
authors = [
    {name = "Infintesky",email = "weiyang2912@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "llvmlite (>=0.44.0,<0.45.0)",
    "numba (>=0.61.0,<0.62.0)",
    "largestinteriorrectangle (>=0.2.1,<0.3.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "edgedetectionfilter", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
